requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
requests==2.31.0
beautifulsoup4==4.12.2
mcp==1.7.1
python-dotenv==1.0.0
Werkzeug==3.0.1
feedparser==6.0.11
pytz==2023.3
tqdm==4.66.1
# Gemini Python SDK for fallback
google-generativeai>=0.7.0,<1.0.0